Moved Logic from FieldRowView.java to FieldRowViewModel.java
